Change Interfaces
Moved all interfaces to their own folder to keep services clean. Made the RunningState enum public and implemented a new interface `IManagedProcess` of which `IAkiServerService` and `ITarkovClientService` inherit to reduce repeated code
