+++ SIT.Manager.Avalonia/App.axaml.cs
@@ -2,6 +2,7 @@
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
 using Microsoft.Extensions.DependencyInjection;
+using SIT.Manager.Avalonia.Interfaces;
 using SIT.Manager.Avalonia.Services;
 using SIT.Manager.Avalonia.ViewModels;
 using SIT.Manager.Avalonia.Views;
