+++ SIT.Manager.Avalonia/Interfaces/IBarNotificationService.cs
@@ -2,7 +2,7 @@
 using SIT.Manager.Avalonia.Models;
 using System;
 
-namespace SIT.Manager.Avalonia.Services
+namespace SIT.Manager.Avalonia.Interfaces
 {
     public interface IBarNotificationService
     {
