+++ SIT.Manager.Avalonia/Interfaces/IFolderPickerService.cs
@@ -1,7 +1,7 @@
 ﻿using Avalonia.Platform.Storage;
 using System.Threading.Tasks;
 
-namespace SIT.Manager.Avalonia.Services
+namespace SIT.Manager.Avalonia.Interfaces
 {
     public interface IFolderPickerService
     {
