+++ SIT.Manager.Avalonia/Interfaces/IVersionService.cs
@@ -1,4 +1,4 @@
-﻿namespace SIT.Manager.Avalonia.Services
+﻿namespace SIT.Manager.Avalonia.Interfaces
 {
     public interface IVersionService
     {
