+++ SIT.Manager.Avalonia/Services/BarNotificationService.cs
@@ -1,4 +1,5 @@
 ﻿using FluentAvalonia.UI.Controls;
+using SIT.Manager.Avalonia.Interfaces;
 using SIT.Manager.Avalonia.Models;
 using System;
