+++ SIT.Manager.Avalonia/Services/ManagerConfigService.cs
@@ -1,4 +1,5 @@
 ﻿using SIT.Manager.Avalonia.Converters;
+using SIT.Manager.Avalonia.Interfaces;
 using SIT.Manager.Avalonia.Models;
 using System;
 using System.Diagnostics;
