+++ SIT.Manager.Avalonia/ViewModels/ServerPageViewModel.cs
@@ -3,8 +3,8 @@ using Avalonia.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using FluentAvalonia.UI.Controls;
+using SIT.Manager.Avalonia.Interfaces;
 using SIT.Manager.Avalonia.Models;
-using SIT.Manager.Avalonia.Services;
 using System;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
